+++ xmake/modules/detect/sdks/find_ndk.lua
@@ -142,9 +142,9 @@ function main(sdkdir, opt)
     opt = opt or {}
 
     -- attempt to load cache first
-    local key = "detect.sdks.find_ndk." .. (sdkdir or "")
+    local key = "detect.sdks.find_ndk"
     local cacheinfo = cache.load(key)
-    if not opt.force and cacheinfo.ndk then
+    if not opt.force and cacheinfo.ndk and cacheinfo.ndk.sdkdir and os.isdir(cacheinfo.ndk.sdkdir) then
         return cacheinfo.ndk
     end
 
@@ -152,7 +152,7 @@ function main(sdkdir, opt)
     local arch = opt.arch or config.get("arch") or "armv7-a"
        
     -- find ndk
-    local ndk = _find_ndk(sdkdir or config.get("ndk") or global.get("ndk"), arch, opt.sdkver or config.get("ndk_sdkver"), opt.toolchains_ver or config.get("ndk_toolchains_ver"))
+    local ndk = _find_ndk(sdkdir or config.get("ndk") or global.get("ndk") or config.get("sdk"), arch, opt.sdkver or config.get("ndk_sdkver"), opt.toolchains_ver or config.get("ndk_toolchains_ver"))
     if ndk and ndk.sdkdir then
 
         -- save to config